# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: Zawadi — Premium Kenyan Corporate Gifting Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /collections:
    get:
      operationId: listCollections
      tags: [collections]
      summary: List curated gift collections
      parameters:
        - name: occasion
          in: query
          schema: { type: string }
        - name: is_featured
          in: query
          schema: { type: boolean }
        - name: min_price
          in: query
          schema: { type: number }
        - name: max_price
          in: query
          schema: { type: number }
      responses:
        "200":
          description: List of collections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Collection"
    post:
      operationId: createCollection
      tags: [collections]
      summary: Create a curated collection (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCollectionInput"
      responses:
        "201":
          description: Created collection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"
  /collections/{id}:
    get:
      operationId: getCollection
      tags: [collections]
      summary: Get collection details with products
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Collection with products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionWithProducts"
    put:
      operationId: updateCollection
      tags: [collections]
      summary: Update a collection
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCollectionInput"
      responses:
        "200":
          description: Updated collection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"

  # ── Hampers ───────────────────────────────────────────────────────────────
...
  /corporates:
    get:
      operationId: listCorporates
      tags: [corporates]
      summary: List corporate accounts (admin)
      parameters:
        - name: tier
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
      responses:
        "200":
          description: List of corporates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Corporate"
    post:
      operationId: createCorporate
      tags: [corporates]
      summary: Register a new corporate account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCorporateInput"
      responses:
        "201":
          description: Created corporate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Corporate"
  /corporates/{id}:
    get:
      operationId: getCorporate
      tags: [corporates]
      summary: Get corporate account details
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Corporate details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Corporate"
    put:
      operationId: updateCorporate
      tags: [corporates]
      summary: Update corporate account
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCorporateInput"
      responses:
        "200":
          description: Updated corporate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Corporate"

  # ── Quotes ────────────────────────────────────────────────────────────────
...
  /dashboard/corporate/{corporateId}:
    get:
      operationId: getCorporateDashboard
      tags: [dashboard]
      summary: Get dashboard data for a specific corporate account
      parameters:
        - name: corporateId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Corporate dashboard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CorporateDashboard"
  /dashboard/recent-orders:
    get:
      operationId: getRecentOrders
      tags: [dashboard]
      summary: Get the most recent orders across all corporates
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 10 }
      responses:
        "200":
          description: Recent orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
  /dashboard/top-products:
    get:
      operationId: getTopProducts
      tags: [dashboard]
      summary: Most ordered products in a date range
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 10 }
      responses:
        "200":
          description: Top products by order volume
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopProduct"

# ── Schemas ──────────────────────────────────────────────────────────────────
components:
  schemas:
...
    Category:
      type: object
      required: [id, name, slug]
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        icon:
          type: string
          nullable: true
        product_count:
          type: integer

    # ── Supplier ──────────────────────────────────────────────────────────
    Supplier:
      type: object
      required: [id, name, county, is_verified, onboarding_status, created_at]
      properties:
        id:
          type: string
        name:
          type: string
        contact_name:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        county:
          type: string
        description:
          type: string
          nullable: true
        story:
          type: string
          nullable: true
        cover_image_url:
          type: string
          nullable: true
        logo_url:
          type: string
          nullable: true
        is_verified:
          type: boolean
        onboarding_status:
          type: string
          enum: [pending, under_review, approved, rejected]
        tags:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
...
    Product:
      type: object
      required: [id, name, category_id, supplier_id, unit_price, moq, lead_time_days, origin, is_active, created_at]
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        category_id:
          type: string
        category:
          $ref: "#/components/schemas/Category"
          nullable: true
        supplier_id:
          type: string
        supplier:
          $ref: "#/components/schemas/Supplier"
          nullable: true
        unit_price:
          type: number
          description: Price in KES per unit at MOQ
        bulk_tiers:
          type: array
          items:
            $ref: "#/components/schemas/BulkPriceTier"
        moq:
          type: integer
          description: Minimum order quantity
        lead_time_days:
          type: integer
        origin:
          type: string
          description: County/region of origin
        tags:
          type: array
          items:
            type: string
        occasion_tags:
          type: array
          items:
            type: string
        images:
          type: array
          items:
            $ref: "#/components/schemas/ProductImage"
        is_active:
          type: boolean
        is_featured:
          type: boolean
        weight_grams:
          type: integer
          nullable: true
        dimensions_cm:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
    BulkPriceTier:
      type: object
      required: [min_qty, price_per_unit]
      properties:
        min_qty:
          type: integer
        price_per_unit:
          type: number
    ProductImage:
      type: object
      required: [id, url, is_primary]
      properties:
        id:
          type: string
        url:
          type: string
        alt:
          type: string
          nullable: true
        is_primary:
          type: boolean
...
    Collection:
      type: object
      required: [id, name, slug, occasion, is_featured, min_price, created_at]
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        occasion:
          type: string
          enum: [client_gifts, staff_appreciation, event_giveaways, festive_hampers, onboarding_kits, custom]
        description:
          type: string
          nullable: true
        cover_image_url:
          type: string
          nullable: true
        is_featured:
          type: boolean
        min_price:
          type: number
        max_price:
          type: number
          nullable: true
        product_count:
          type: integer
        created_at:
          type: string
          format: date-time
    CollectionWithProducts:
      allOf:
        - $ref: "#/components/schemas/Collection"
        - type: object
          properties:
            products:
              type: array
              items:
                $ref: "#/components/schemas/Product"
    CreateCollectionInput:
      type: object
      required: [name, occasion]
      properties:
        name:
          type: string
        occasion:
          type: string
          enum: [client_gifts, staff_appreciation, event_giveaways, festive_hampers, onboarding_kits, custom]
        description:
          type: string
          nullable: true
        cover_image_url:
          type: string
          nullable: true
        is_featured:
          type: boolean
        product_ids:
          type: array
          items:
            type: string
    UpdateCollectionInput:
      type: object
      properties:
        name:
          type: string
        occasion:
          type: string
        description:
          type: string
          nullable: true
        cover_image_url:
          type: string
          nullable: true
        is_featured:
          type: boolean
        product_ids:
          type: array
          items:
            type: string

    # ── Hamper ────────────────────────────────────────────────────────────
...
    Corporate:
      type: object
      required: [id, name, email, tier, created_at]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        kra_pin:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        industry:
          type: string
          nullable: true
        tier:
          type: string
          enum: [standard, premium, enterprise]
        payment_terms:
          type: string
          enum: [prepaid, net_15, net_30, net_60]
          nullable: true
        account_manager_name:
          type: string
          nullable: true
        total_orders:
          type: integer
        total_spend:
          type: number
        created_at:
          type: string
          format: date-time
    CreateCorporateInput:
      type: object
      required: [name, email]
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        kra_pin:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        industry:
          type: string
          nullable: true
    UpdateCorporateInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        kra_pin:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        industry:
          type: string
          nullable: true
        tier:
          type: string
          enum: [standard, premium, enterprise]
        payment_terms:
          type: string
          enum: [prepaid, net_15, net_30, net_60]
          nullable: true
        account_manager_name:
          type: string
          nullable: true

    # ── Quote ─────────────────────────────────────────────────────────────
...
    Order:
      type: object
      required: [id, corporate_id, status, subtotal, vat, total, created_at]
      properties:
        id:
          type: string
        reference:
          type: string
          description: Human-readable order reference e.g. ZWD-2026-0042
        corporate_id:
          type: string
        corporate_name:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, confirmed, in_production, quality_check, packaging, dispatched, delivered, cancelled]
        subtotal:
          type: number
        vat:
          type: number
        total:
          type: number
        delivery_address:
          type: string
          nullable: true
        delivery_date:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
          nullable: true
        recipient_count:
          type: integer
        created_at:
          type: string
          format: date-time
...
    CorporateDashboard:
      type: object
      required: [corporate_id, total_orders, total_spend, active_orders]
      properties:
        corporate_id:
          type: string
        total_orders:
          type: integer
        total_spend:
          type: number
        active_orders:
          type: integer
        recent_orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        upcoming_occasions:
          type: array
          items:
            type: object
            properties:
              occasion:
                type: string
              date:
                type: string
              days_until:
                type: integer
        spend_by_month:
          type: array
          items:
            type: object
            required: [month, spend]
            properties:
              month:
                type: string
              spend:
                type: number
    TopProduct:
      type: object
      required: [product_id, product_name, total_units, total_revenue]
      properties:
        product_id:
          type: string
        product_name:
          type: string
        supplier_name:
          type: string
          nullable: true
        total_units:
          type: integer
        total_revenue:
          type: number
        image_url:
          type: string
          nullable: true