# lib/api-zod/src/generated/types/collectionWithProducts.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Zawadi — Premium Kenyan Corporate Gifting Platform API
 * OpenAPI spec version: 1.0.0
 */
import type { Collection } from "./collection";
import type { Product } from "./product";

export type CollectionWithProducts = Collection & {
  products?: Product[];
};