# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Zawadi — Premium Kenyan Corporate Gifting Platform API
 * OpenAPI spec version: 1.0.0
 */
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: string;
  /** Human-readable order reference e.g. ZWD-2026-0042 */
  reference?: string;
  corporate_id: string;
  corporate_name?: string | null;
  status: OrderStatus;
  subtotal: number;
  vat: number;
  total: number;
  delivery_address?: string | null;
  delivery_date?: Date | null;
  notes?: string | null;
  recipient_count?: number;
  created_at: Date;
}